+++ src/StudioCore/Editor/Action.cs
@@ -140,222 +140,6 @@ public class PropertiesChangedAction : EditorAction
         public PropertyInfo Property;
     }
 }
-
-public class AddParamsAction : EditorAction
-{
-    private readonly bool appOnly;
-    private readonly List<Param.Row> Clonables = new();
-    private readonly List<Param.Row> Clones = new();
-    private readonly int InsertIndex;
-    private readonly Param Param;
-    private readonly List<Param.Row> Removed = new();
-    private readonly List<int> RemovedIndex = new();
-    private readonly bool replParams;
-    private string ParamString;
-
-    public AddParamsAction(Param param, string pstring, List<Param.Row> rows, bool appendOnly, bool replaceParams,
-        int index = -1)
-    {
-        Param = param;
-        Clonables.AddRange(rows);
-        ParamString = pstring;
-        appOnly = appendOnly;
-        replParams = replaceParams;
-        InsertIndex = index;
-    }
-
-    public override ActionEvent Execute()
-    {
-        foreach (Param.Row row in Clonables)
-        {
-            var newrow = new Param.Row(row);
-            if (InsertIndex > -1)
-            {
-                newrow.Name = row.Name != null ? row.Name + "_1" : "";
-                Param.InsertRow(InsertIndex, newrow);
-            }
-            else
-            {
-                if (Param[row.ID] != null)
-                {
-                    if (replParams)
-                    {
-                        Param.Row existing = Param[row.ID];
-                        RemovedIndex.Add(Param.IndexOfRow(existing));
-                        Removed.Add(existing);
-                        Param.RemoveRow(existing);
-                    }
-                    else
-                    {
-                        newrow.Name = row.Name != null ? row.Name + "_1" : "";
-                        var newID = row.ID + 1;
-                        while (Param[newID] != null)
-                        {
-                            newID++;
-                        }
-
-                        newrow.ID = newID;
-                        Param.InsertRow(Param.IndexOfRow(Param[newID - 1]) + 1, newrow);
-                    }
-                }
-
-                if (Param[row.ID] == null)
-                {
-                    newrow.Name = row.Name != null ? row.Name : "";
-                    if (appOnly)
-                    {
-                        Param.AddRow(newrow);
-                    }
-                    else
-                    {
-                        var index = 0;
-                        foreach (Param.Row r in Param.Rows)
-                        {
-                            if (r.ID > newrow.ID)
-                            {
-                                break;
-                            }
-
-                            index++;
-                        }
-
-                        Param.InsertRow(index, newrow);
-                    }
-                }
-            }
-
-            Clones.Add(newrow);
-        }
-
-        // Refresh diff cache
-        TaskManager.Run(new TaskManager.LiveTask("Param - Check Differences",
-            TaskManager.RequeueType.Repeat, true,
-            TaskLogs.LogPriority.Low,
-            () => ParamDiffBank.RefreshAllParamDiffCaches(false)));
-        return ActionEvent.NoEvent;
-    }
-
-    public override ActionEvent Undo()
-    {
-        for (var i = 0; i < Clones.Count(); i++)
-        {
-            Param.RemoveRow(Clones[i]);
-        }
-
-        for (var i = Removed.Count() - 1; i >= 0; i--)
-        {
-            Param.InsertRow(RemovedIndex[i], Removed[i]);
-        }
-
-        Clones.Clear();
-        RemovedIndex.Clear();
-        Removed.Clear();
-        return ActionEvent.NoEvent;
-    }
-
-    public List<Param.Row> GetResultantRows()
-    {
-        return Clones;
-    }
-}
-
-public class DeleteParamsAction : EditorAction
-{
-    private readonly List<Param.Row> Deletables = new();
-    private readonly Param Param;
-    private readonly List<int> RemoveIndices = new();
-    private readonly bool SetSelection = false;
-
-    public DeleteParamsAction(Param param, List<Param.Row> rows)
-    {
-        Param = param;
-        Deletables.AddRange(rows);
-    }
-
-    public override ActionEvent Execute()
-    {
-        foreach (Param.Row row in Deletables)
-        {
-            RemoveIndices.Add(Param.IndexOfRow(row));
-            Param.RemoveRowAt(RemoveIndices.Last());
-        }
-
-        if (SetSelection)
-        {
-        }
-
-        return ActionEvent.NoEvent;
-    }
-
-    public override ActionEvent Undo()
-    {
-        for (var i = Deletables.Count() - 1; i >= 0; i--)
-        {
-            Param.InsertRow(RemoveIndices[i], Deletables[i]);
-        }
-
-        if (SetSelection)
-        {
-        }
-
-        // Refresh diff cache
-        TaskManager.Run(new TaskManager.LiveTask("Param - Check Differences",
-            TaskManager.RequeueType.Repeat, true,
-            TaskLogs.LogPriority.Low,
-            () => ParamDiffBank.RefreshAllParamDiffCaches(false)));
-        return ActionEvent.NoEvent;
-    }
-}
-
-public class DuplicateFMGEntryAction : EditorAction
-{
-    private readonly FMGEntryGroup EntryGroup;
-    private FMGEntryGroup NewEntryGroup;
-
-    public DuplicateFMGEntryAction(FMGEntryGroup entryGroup)
-    {
-        EntryGroup = entryGroup;
-    }
-
-    public override ActionEvent Execute()
-    {
-        NewEntryGroup = EntryGroup.DuplicateFMGEntries();
-        NewEntryGroup.SetNextUnusedID();
-        return ActionEvent.NoEvent;
-    }
-
-    public override ActionEvent Undo()
-    {
-        NewEntryGroup.DeleteEntries();
-        return ActionEvent.NoEvent;
-    }
-}
-
-public class DeleteFMGEntryAction : EditorAction
-{
-    private FMGEntryGroup BackupEntryGroup = new();
-    private FMGEntryGroup EntryGroup;
-
-    public DeleteFMGEntryAction(FMGEntryGroup entryGroup)
-    {
-        EntryGroup = entryGroup;
-    }
-
-    public override ActionEvent Execute()
-    {
-        BackupEntryGroup = EntryGroup.CloneEntryGroup();
-        EntryGroup.DeleteEntries();
-        return ActionEvent.NoEvent;
-    }
-
-    public override ActionEvent Undo()
-    {
-        EntryGroup = BackupEntryGroup;
-        EntryGroup.ImplementEntryGroup();
-        return ActionEvent.NoEvent;
-    }
-}
-
 public class CompoundAction : EditorAction
 {
     private readonly List<EditorAction> Actions;
